ofa_cookiecontrol.js - One Script For All
This script refers to the one that will handle all cookies among all Joomla! properties. This is the outcome of Compliance Team's Cookie Audit and as a result we have this release as the beta version. Also there is an implementation of the GTM script in a way that every property can parse the GTM's id as variable inside the script. Also, every similar script, based on the category that belongs, could be implemented on the correspondent onAccept function.